np float32 w/transformer

+++ autograd/cpu/constants.py
@@ -1,6 +1,6 @@
 
 # define what data type you want
 import numpy as np
-PRECISION = np.float64
+PRECISION = np.float32
 # one of np.float32, np.float64, np.float128
 # if not using softmax, can use np.float16
